# openapi.yaml
openapi: 3.1.0
info:
    title: VoiceBlender API
    description: |
        VoiceBlender bridges SIP and WebRTC voice calls with multi-party audio mixing, real-time speech-to-text, text-to-speech, AI agent integration, recording, and webhook-based event delivery.
    x-config-vars:
        - name: INSTANCE_ID
          default: (auto-generated UUID)
          description: Instance identifier included in all API responses and webhook events
        - name: HTTP_ADDR
          default: :8080
          description: REST API listen address
        - name: SIP_BIND_IP
          default: 127.0.0.1
          description: IPv4 address advertised in SDP, Contact, and Via headers (and used as the listen address when SIP_LISTEN_IP is empty)
        - name: SIP_LISTEN_IP
          default: (same as SIP_BIND_IP)
          description: UDP socket bind IP. Accepts 127.0.0.1, 0.0.0.0, ::, or any literal v4/v6 address
        - name: SIP_BIND_IPV6
          default: (empty = v4-only)
          description: IPv6 address advertised in SDP/Contact/Via for IPv6 calls. Set this for IPv6-only or dual-stack deployments
        - name: SIP_LISTEN_IPV6
          default: (same as SIP_BIND_IPV6)
          description: Optional separate IPv6 socket bind address (used when running with both 0.0.0.0 and a specific v6 literal)
        - name: SIP_PORT
          default: "5060"
          description: SIP listen port
        - name: SIP_HOST
          default: voiceblender
...
        - name: ICE_SERVERS
...
servers:
...
paths:
    /legs:
        post:
            operationId: createLeg
            summary: Originate an outbound SIP call
            tags:
                - Legs
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateLegRequest'
            responses:
                '201':
                    description: Leg created
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Leg'
                '400':
                    description: Invalid JSON, bad SIP URI, unknown codec, or unsupported type
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
        get:
            operationId: listLegs
            summary: List all active legs
            tags:
                - Legs
            responses:
                '200':
                    description: Array of legs
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: '#/components/schemas/Leg'
    /legs/{id}:
        parameters:
            - $ref: '#/components/parameters/LegId'
        get:
            operationId: getLeg
            summary: Get a single leg
            tags:
                - Legs
            responses:
                '200':
                    description: Leg details
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Leg'
                '404':
                    description: Leg not found
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
        delete:
            operationId: deleteLeg
            summary: Hang up a leg (asynchronous)
            description: |-
                Validates the leg exists and queues a hangup. The HTTP call returns 202 as soon as the leg is found; the SIP work and cleanup run in the background, and the eventual disconnection is observed via the `leg.disconnected` event.

                Without a request body the legacy behavior is preserved: SIP BYE on connected legs (`cdr.reason: "api_hangup"`), or dialog cancel on unanswered inbound legs (`cdr.reason: "caller_cancel"`).

                With `{"reason": "<value>"}` and an unanswered SIP inbound leg (state `ringing` or `early_media`), VoiceBlender sends a final non-2xx response instead of BYE/cancel: `busy`→486, `declined`/`rejected`→603, `unavailable`→480, `not_found`→404, `forbidden`→403, `server_error`→500. The reason value is passed through to `leg.disconnected`'s `cdr.reason`.

                For connected legs the request body is ignored.
            tags:
                - Legs
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/DeleteLegRequest'
            responses:
                '202':
                    description: Hangup queued
                '400':
                    description: Unknown reason value
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                '404':
                    description: Leg not found
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
    /legs/{id}/answer:
        parameters:
            - $ref: '#/components/parameters/LegId'
        post:
            operationId: answerLeg
            summary: Answer a ringing or early-media inbound SIP leg (asynchronous)
            description: Signals the inbound-call goroutine to send 200 OK. The HTTP call returns 202 immediately; the actual SIP 200 OK is sent in the background, and the leg's transition is observed via `leg.connected`. Pre-condition failures (wrong state, unknown codec) still return 4xx synchronously.
            tags:
                - Legs
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/AnswerLegRequest'
            responses:
                '202':
                    description: Answer queued
                '400':
                    description: Not a SIP inbound leg, invalid body, or codec not in offer
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                '404':
                    description: Leg not found
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                '409':
                    description: Leg is not in ringing or early_media state
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
    /legs/{id}/ring:
        parameters:
            - $ref: '#/components/parameters/LegId'
        post:
            operationId: ringLeg
            summary: Send 180 Ringing on a ringing inbound SIP leg (asynchronous)
            description: 'Queues a SIP 180 Ringing provisional response with no SDP. Use when `SIP_AUTO_RINGING=false` (the default) and you want to indicate alerting before deciding to early-media or answer. Idempotent: each call emits another 180 — receivers tolerate re-sends. The HTTP call returns 202 as soon as the request is validated; SIP-level send failures surface as `leg.command_failed` with `command="ring"`.'
            tags:
                - Legs
            responses:
                '202':
                    description: 180 Ringing queued
                '400':
                    description: Not a SIP inbound leg
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                '404':
                    description: Leg not found
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                '409':
                    description: Leg is not in ringing state
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
    /legs/{id}/early-media:
        parameters:
            - $ref: '#/components/parameters/LegId'
        post:
            operationId: earlyMediaLeg
            summary: Enable early media on a ringing inbound SIP leg (asynchronous)
            description: Queues a SIP 183 Session Progress with SDP and the RTP/codec setup. The HTTP call returns 202 as soon as the request is validated; the leg transitions to `early_media` state asynchronously, observable via `leg.early_media`. Setup failures surface as `leg.command_failed` with `command="early_media"`.
            tags:
                - Legs
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/EarlyMediaLegRequest'
            responses:
                '202':
                    description: Early media queued
                '400':
                    description: Not a SIP inbound leg or codec not in offer
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                '404':
                    description: Leg not found
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                '409':
                    description: Leg is not in ringing state
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
    /legs/{id}/mute:
        parameters:
            - $ref: '#/components/parameters/LegId'
        post:
            operationId: muteLeg
            summary: Mute a leg
            description: A muted leg's audio is excluded from the room mix and speaking events are suppressed. Taps (recording/STT) still receive the muted leg's own audio.
            tags:
                - Legs
            responses:
                '200':
                    description: Leg muted
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/StatusResponse'
                '404':
                    description: Leg not found
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
        delete:
            operationId: unmuteLeg
            summary: Unmute a leg
            tags:
                - Legs
            responses:
                '200':
                    description: Leg unmuted
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/StatusResponse'
                '404':
                    description: Leg not found
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
    /legs/{id}/hold:
        parameters:
            - $ref: '#/components/parameters/LegId'
        post:
            operationId: holdLeg
            summary: Put a SIP call on hold (asynchronous)
            description: Queues a re-INVITE with `sendonly` SDP direction. The HTTP call returns 202 as soon as the leg is validated; the re-INVITE is sent in the background and success surfaces as `leg.hold`. Failures surface as `leg.command_failed` with `command="hold"`. The RTP timeout is paused while held, and a 2-hour auto-hangup timer starts.
            tags:
                - Legs
            responses:
                '202':
                    description: Hold queued
                '400':
                    description: Not a SIP leg
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                '404':
                    description: Leg not found
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                '409':
                    description: Hold not supported for this leg type (e.g. WhatsApp)
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
        delete:
            operationId: unholdLeg
            summary: Resume a held SIP call (asynchronous)
            description: Queues a re-INVITE with `sendrecv` SDP direction. The HTTP call returns 202; success surfaces as `leg.unhold`, failures as `leg.command_failed` with `command="unhold"`.
            tags:
                - Legs
            responses:
                '202':
                    description: Unhold queued
                '400':
                    description: Not a SIP leg
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                '404':
                    description: Leg not found
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                '409':
                    description: Hold not supported for this leg type (e.g. WhatsApp)
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
    /legs/{id}/transfer:
        parameters:
            - $ref: '#/components/parameters/LegId'
        post:
            operationId: transferLeg
            summary: Transfer a SIP leg via REFER (asynchronous)
            description: Asynchronously transfers a SIP leg. The HTTP call returns 202 as soon as the request is validated; the REFER is sent in the background and its outcome is surfaced through `leg.transfer_initiated` / `leg.transfer_progress` / `leg.transfer_completed` / `leg.transfer_failed` events. Blind transfer when `replaces_leg_id` is omitted; attended transfer when present (the named leg's dialog identity is embedded as a Replaces parameter per RFC 3891). On terminal 2xx the leg (and the replaces leg, if any) is hung up automatically.
            tags:
                - Legs
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/TransferRequest'
            responses:
                '202':
                    description: Transfer request accepted for processing
                '400':
                    description: Missing or invalid target URI (including URIs without a host such as sip:)
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                '404':
                    description: Leg not found
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                '409':
                    description: Leg not connected, not a SIP leg, or replaces_leg_id is invalid
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
    /legs/{id}/dtmf:
        parameters:
            - $ref: '#/components/parameters/LegId'
        post:
            operationId: sendDTMF
            summary: Send DTMF digits on a leg
            tags:
                - Legs
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/DTMFRequest'
            responses:
                '200':
                    description: Digits sent
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/StatusResponse'
                '400':
                    description: Invalid JSON or empty digits
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                '404':
                    description: Leg not found
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                '500':
                    description: DTMF writer unavailable
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
    /legs/{id}/dtmf/accept:
        parameters:
            - $ref: '#/components/parameters/LegId'
        post:
            operationId: acceptDTMFLeg
            summary: Enable DTMF reception on a leg
            description: Allow this leg to receive DTMF digits broadcast from other legs in the same room. This is the default state for new legs.
            tags:
                - Legs
            responses:
                '200':
                    description: DTMF reception enabled
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/StatusResponse'
                '404':
                    description: Leg not found
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
    /legs/{id}/dtmf/reject:
        parameters:
            - $ref: '#/components/parameters/LegId'
        post:
            operationId: rejectDTMFLeg
            summary: Disable DTMF reception on a leg
            description: Block this leg from receiving DTMF digits broadcast from other legs in the same room. DTMF received from this leg's own far end is still emitted as a leg.dtmf event.
            tags:
                - Legs
            responses:
                '200':
                    description: DTMF reception disabled
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/StatusResponse'
                '404':
                    description: Leg not found
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
    /legs/{id}/rtt:
        parameters:
            - $ref: '#/components/parameters/LegId'
        post:
            operationId: sendRTT
            summary: Send Real-Time Text (T.140) on a SIP leg
            description: Sends UTF-8 text on the leg's RTT (T.140 / RFC 4103) media stream. Requires that the SDP offer/answer agreed on an m=text section with the remote UA. Enable RTT on the server with RTT_ENABLED=true.
            tags:
                - Legs
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/RTTRequest'
            responses:
                '200':
                    description: Text sent
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/StatusResponse'
                '400':
                    description: Invalid JSON or empty text
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                '404':
                    description: Leg not found
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                '409':
                    description: RTT was not negotiated for this leg
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                '500':
                    description: Send failed
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
    /legs/{id}/rtt/accept:
        parameters:
            - $ref: '#/components/parameters/LegId'
        post:
            operationId: acceptRTTLeg
            summary: Enable RTT reception on a leg
            description: Allow this leg to receive RTT text broadcast from other legs in the same room and to emit rtt.received events for incoming text. Default state for new legs.
            tags:
                - Legs
            responses:
                '200':
                    description: RTT reception enabled
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/StatusResponse'
                '404':
                    description: Leg not found
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
    /legs/{id}/rtt/reject:
        parameters:
            - $ref: '#/components/parameters/LegId'
        post:
            operationId: rejectRTTLeg
            summary: Disable RTT reception on a leg
            description: Block this leg from receiving RTT text broadcast from other legs in the same room and suppress rtt.received events for this leg.
            tags:
                - Legs
            responses:
                '200':
                    description: RTT reception disabled
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/StatusResponse'
                '404':
                    description: Leg not found
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
    /legs/{id}/play:
        parameters:
            - $ref: '#/components/parameters/LegId'
        post:
            operationId: playLeg
            summary: Start audio playback to a leg
            tags:
                - Legs
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/PlaybackRequest'
            responses:
                '200':
                    description: Playback started
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/StatusResponse'
                '400':
                    description: Invalid JSON or volume out of range
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                '404':
                    description: Leg not found
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                '409':
                    description: Leg has no audio writer
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
    /legs/{id}/play/{playbackID}:
        parameters:
            - $ref: '#/components/parameters/LegId'
            - $ref: '#/components/parameters/PlaybackId'
        patch:
            operationId: volumePlayLeg
            summary: Change the volume of an active leg playback
            tags:
                - Legs
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/VolumeRequest'
            responses:
                '200':
                    description: Volume updated
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/StatusResponse'
                '400':
                    description: Invalid JSON or volume out of range
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                '404':
                    description: Playback not found
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
        delete:
            operationId: stopPlayLeg
            summary: Stop audio playback on a leg
            tags:
                - Legs
            responses:
                '200':
                    description: Playback stopped
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/StatusResponse'
                '404':
                    description: No playback in progress
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
...
    /legs/{id}/record/pause:
        parameters:
            - $ref: '#/components/parameters/LegId'
        post:
            operationId: pauseRecordLeg
            summary: Pause a leg recording
            description: 'Replaces incoming audio with silence on the active recording until `/record/resume` is called. The WAV''s timeline is preserved (silent gap where audio was paused), so reviewers can see exactly when sensitive data was excluded. Idempotent: calling while already paused returns `status: already_paused`.'
            tags:
                - Legs
            responses:
                '200':
                    description: Recording paused (or already paused)
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/StatusResponse'
                '404':
                    description: No recording in progress
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
...
    /legs/{id}/agent:
        parameters:
            - $ref: '#/components/parameters/LegId'
        delete:
            operationId: stopAgentLeg
            summary: Detach the agent from a leg
            tags:
                - Legs
            responses:
                '200':
                    description: Agent stopped
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/StatusResponse'
                '404':
                    description: No agent attached to this leg
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
...
components:
...
    schemas:
...
        Error:
            type: object
            properties:
                instance_id:
                    type: string
                    description: Instance identifier
                error:
                    type: string
                    description: Error message
            required:
                - error
...